backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
tqdm==4.66.5

torch==2.4.1
transformers==4.44.2
sentencepiece==0.2.0
accelerate==0.34.2

faiss-cpu==1.8.0.post1
chromadb==0.5.5

spacy==3.7.6
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

sqlalchemy==2.0.35
psycopg2-binary==2.9.9

ollama==0.3.3
langchain==0.3.1

python-dotenv==1.0.1
orjson==3.10.7
httpx==0.27.2
